+++ src/app/store/actions/node.actions.ts
@@ -37,6 +37,7 @@ export const EDIT_FOLDER = 'EDIT_FOLDER';
 export const SHARE_NODE = 'SHARE_NODE';
 export const UNSHARE_NODES = 'UNSHARE_NODES';
 export const COPY_NODES = 'COPY_NODES';
+export const MOVE_NODES = 'MOVE_NODES';
 
 export class SetSelectedNodesAction implements Action {
     readonly type = SET_SELECTED_NODES;
@@ -90,5 +91,10 @@ export class UnshareNodesAction implements Action {
 
 export class CopyNodesAction implements Action {
     readonly type = COPY_NODES;
+    constructor(public payload: Array<MinimalNodeEntity>) {}
+}
+
+export class MoveNodesAction implements Action {
+    readonly type = MOVE_NODES;
     constructor(public payload: Array<MinimalNodeEntity>) {}
 }
